Move logic for handling Tendermint validator updates to validator_set.rs
